Update DiT-MoE documentation and pipeline configurations
- Marked DiT-MoE as completed in the README checklist for better tracking.
- Updated README and example code to reflect the change from `torch.float16` to `torch.bfloat16` for improved performance.
- Added licensing information to the DiT-MoE pipeline files for compliance and clarity.

+++ src/diffusers/DiT-MoE/pipeline.py
@@ -4,6 +4,20 @@ Load with native Hugging Face diffusers and trust_remote_code=True.
 
 from __future__ import annotations
 
+# Copyright 2026 The HuggingFace Team. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import importlib
 import inspect
 import json
@@ -34,7 +48,7 @@ EXAMPLE_DOC_STRING = """
         ...     local_files_only=True,
         ...     custom_pipeline=str(model_dir / "pipeline.py"),
         ...     trust_remote_code=True,
-        ...     torch_dtype=torch.float16,
+        ...     torch_dtype=torch.bfloat16,
         ... )
         >>> pipe.to("cuda")
